Replaces the static field with AsyncLocal and simplifies the connection management.
