+++ samples/outbox/multi-tenant/Version_6/Receiver/MultiTenantConnectionProvider.cs
@@ -1,30 +1,23 @@
-﻿using System;
-using System.Configuration;
-using System.Data;
+﻿using System.Data;
 using NHibernate.Connection;
-using Receiver;
 
 class MultiTenantConnectionProvider : DriverConnectionProvider
 {
     public override IDbConnection GetConnection()
     {
-        string defaultConnectionString = ConfigurationManager.ConnectionStrings["NServiceBus/Persistence"]
-            .ConnectionString;
-
         #region GetConnectionFromContext
 
-        Lazy<IDbConnection> lazy;
-        PipelineExecution pipelineExecutor = PipelineExecution.Instance;
-        string key = $"LazySqlConnection-{defaultConnectionString}";
-        if (pipelineExecutor.CurrentContext != null && pipelineExecutor.CurrentContext.Extensions.TryGet(key, out lazy))
+        string connectionString = ExtractTenantConnectionStringBehavior.ConnectionStringHolder.Value;
+        if (connectionString != null)
         {
             IDbConnection connection = Driver.CreateConnection();
-            connection.ConnectionString = lazy.Value.ConnectionString;
+            connection.ConnectionString = connectionString;
             connection.Open();
             return connection;
         }
-        return base.GetConnection();
 
         #endregion
+
+        return base.GetConnection();
     }
 }
