--- samples/outbox/multi-tenant/Version_6/Receiver/PipelineExecution.cs
@@ -1,16 +1,0 @@
-﻿namespace Receiver
-{
-    using NServiceBus.Pipeline;
-
-    public class PipelineExecution
-    {
-        static PipelineExecution()
-        {
-            Instance = new PipelineExecution();
-        }
-
-        public static PipelineExecution Instance { get; }
-
-        public IIncomingPhysicalMessageContext CurrentContext { get; set; }
-    }
-}
