+++ samples/outbox/multi-tenant/Version_6/Receiver/Receiver.csproj
@@ -9,7 +9,7 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>Receiver</RootNamespace>
     <AssemblyName>Receiver</AssemblyName>
-    <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
   </PropertyGroup>
@@ -48,8 +48,9 @@
     <Reference Include="NServiceBus.Core">
       <HintPath>..\..\..\..\..\packages\NServiceBus.6.0.0-beta0001\lib\net452\NServiceBus.Core.dll</HintPath>
     </Reference>
-    <Reference Include="NServiceBus.NHibernate">
-      <HintPath>..\..\..\..\..\packages\NServiceBus.NHibernate.7.0.0-unstable0073\lib\net452\NServiceBus.NHibernate.dll</HintPath>
+    <Reference Include="NServiceBus.NHibernate, Version=7.0.0.0, Culture=neutral, PublicKeyToken=9fc386479f8a226c, processorArchitecture=MSIL">
+      <HintPath>..\..\..\..\..\packages\NServiceBus.NHibernate.7.0.0-beta0001\lib\net452\NServiceBus.NHibernate.dll</HintPath>
+      <Private>True</Private>
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Configuration" />
@@ -63,14 +64,13 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="MultiTenantConnectionProvider.cs" />
-    <Compile Include="MultiTenantOpenSqlConnectionBehavior.cs" />
+    <Compile Include="ExtractTenantConnectionStringBehavior.cs" />
     <Compile Include="Order.cs" />
     <Compile Include="OrderLifecycleSaga.cs" />
     <Compile Include="OrderLifecycleSagaData.cs" />
     <Compile Include="OrderMap.cs" />
     <Compile Include="OrderSubmittedHandler.cs" />
     <Compile Include="OrderTimeout.cs" />
-    <Compile Include="PipelineExecution.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="PropagateIncomingTenantIdBehavior.cs" />
     <Compile Include="PropagateOutgoingTenantIdBehavior.cs" />
