+++ samples/outbox/multi-tenant/Version_6/Sender/packages.config
@@ -3,5 +3,5 @@
   <package id="Iesi.Collections" version="4.0.1.4000" targetFramework="net452" />
   <package id="NHibernate" version="4.0.4.4000" targetFramework="net452" />
   <package id="NServiceBus" version="6.0.0-beta0001" targetFramework="net452" />
-  <package id="NServiceBus.NHibernate" version="7.0.0-unstable0073" targetFramework="net452" />
+  <package id="NServiceBus.NHibernate" version="7.0.0-beta0001" targetFramework="net452" />
 </packages>
